part1/requirements_fixed.txt
# Core ML dependencies
pandas>=1.3.0
scikit-learn==1.0.2
numpy>=1.21.0
joblib>=1.0.0

# API and web
requests==2.25.1
flask==2.0.1
gunicorn==20.1.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0

# Database
psycopg2-binary==2.8.6
sqlalchemy>=1.4.0
alembic>=1.7.0

# Testing
pytest>=6.0.0
pytest-cov==2.12.1
flake8>=3.9.0
black>=21.0.0
isort>=5.9.0
pre-commit>=2.15.0

# Visualization
matplotlib>=3.3.0
seaborn>=0.11.0

# Configuration
PyYAML>=5.4.0
click>=8.0.0

# Monitoring
prometheus-client>=0.11.0

# Optional dependencies (commented out to reduce conflicts)
# tensorflow==2.6.0  # Only needed for deep learning models
# torch>=1.9.0       # Only needed for deep learning models
# torchvision>=0.10.0  # Only needed for deep learning models

# Cloud and orchestration
boto3>=1.18.0
# kubernetes>=18.0.0  # Only needed for K8s deployments
# docker              # Only needed for container management
# mlflow==1.20.0      # Only needed for experiment tracking
# redis>=3.5.3        # Only needed for caching
# celery[redis]>=5.1.2  # Only needed for task queuing